refactor: user -> movie 이동
기존의 userPreferredGenreMovieList(), userWatchingMovieList()를
movie로 이동하며 관련 파일들도 movie로 이동했습니다.
